% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{explainHashDataFrame}
\alias{explainHashDataFrame}
\title{Provides explanation, how given data frame will be hashed.}
\usage{
explainHashDataFrame(df, numCols, hashSeed, factorMode = "integer")
}
\arguments{
\item{df}{Data frame to be hashed.}

\item{numCols}{Integer, number of columns for the output data frame.}

\item{hashSeed}{Integer, seed for the employed hash function (MurmurHash3).}

\item{factorMode, }{determines how factors will be hashed --- \code{"integer"},
treats factors as normal integers, \code{"oneHot"} uses one hot encoding,
\code{"skip"} skips all the factors in the input. Defaults to \code{"integer"}.}
}
\value{
List of integer vectors. Each entry in list corresponds to one column
in the output of hashing. Entries in integer vectors give signs and names of integer
vector entries provide names of columns in original data frame.
}
\description{
Creates a data structure describing the hashing process.
}

