% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hashDataFrame}
\alias{hashDataFrame}
\title{Hashes the data frame, so the output contains much smaller number of columns.}
\usage{
hashDataFrame(df, numCols, hashSeed)
}
\arguments{
\item{df}{Data frame to be hashed.}

\item{numCols}{Integer, number of columns for the output data frame.}

\item{hashSeed}{Integer, seed for the employed hash function (MurmurHash3).}
}
\value{
Data frame having \code{numCols} columns, containing combination of
columns from \code{df} according to hashing function.
}
\description{
Calculates the reduced representation of the given data frame based on the
hashing trick.
}

