% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hashDataFrame}
\alias{hashDataFrame}
\title{Hashes the data frame \code{df}, so the output contains \code{numCols}}
\usage{
hashDataFrame(df, numCols, hashSeed)
}
\arguments{
\item{df}{data frame to be hashed}

\item{numCols}{integer, number of columns for the output data frame}

\item{hashSeed}{integer, seed for the employed hash function (MurmurHash3)}
}
\description{
Hashes the data frame \code{df}, so the output contains \code{numCols}
}

